back/requirements.txt
aioredis==2.0.1
aiosmtplib==2.0.2
alembic==1.11.3
anyio==3.7.1
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.3
asyncpg==0.28.0
backcall==0.2.0
click==8.1.7
decorator==5.1.1
exceptiongroup==1.1.3
executing==1.2.0
fastapi==0.102.0
fastapi-jwt-auth==0.5.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.19.0
jinja2==3.1.2
mako==1.2.4
markupsafe==2.1.3
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.10.12
pygments==2.16.1
pyjwt==1.7.1
python-multipart==0.0.6
six==1.16.0
sniffio==1.3.0
sqladmin==0.14.1
sqlalchemy==2.0.20
stack-data==0.6.2
starlette==0.27.0
traitlets==5.9.0
typing-extensions==4.7.1
uvicorn==0.23.2
wcwidth==0.2.6
wtforms==3.0.1
aiofiles
pyotp~=2.9.0
cryptography~=41.0.4
paho-mqtt
docopt
numpy
ultralytics
